+++ extensions/easy-new-file/CHANGELOG.md
@@ -1,7 +1,9 @@
 # Easy New File Changelog
 
 ## [Added Easy New File] - 2022-04-15
+
 - Release of New File Here version 1.0
 - **Create File Here**: Create a new file in the current Finder window or on the desktop.
   - Create File with Name: Create a new file with name.
+  - Copy File to Clipboard: Create the file to clipboard.
 - **Add File Template**: Add a file template that you can use in the "New File Here" command.